refactor: 해야할 일들 정리하기
