+++ src/main/kotlin/blackjack/domain/card/CardsState.kt
@@ -1,5 +1,6 @@
 package blackjack.domain.card
 
+//TODO: 왜 내가 sealed class를 사용했을까?
 sealed class CardsState {
 
     object BlackJack : CardsState()
